fix(vscode): align session cost with usage
