+++ .changeset/tui-scoped-session-cost.md
@@ -1,5 +1,6 @@
 ---
 "@kilocode/cli": patch
+"kilo-code": patch
 ---
 
-Keep TUI session costs consistent and scope subagent usage to the viewed session tree.
+Keep TUI and VS Code session costs consistent and scope subagent usage to the viewed session tree.
