+++ packages/kilo-vscode/webview-ui/src/context/model-usage.ts
@@ -4,6 +4,17 @@ const DATE_SUFFIX = /(?:-(?:20\d{6}|20\d{2}-\d{2}-\d{2}))(?:-v\d+(?::\d+)?)?$/i
 
 export type TokenSummary = { input: number; output: number; cached: number }
 
+export function formatCost(input: number, locale: string) {
+  const value = Math.max(0, Number.isFinite(input) ? input : 0)
+  if (value > 0 && value < 0.000001) return "<$0.000001"
+  return new Intl.NumberFormat(locale, {
+    style: "currency",
+    currency: "USD",
+    minimumFractionDigits: 2,
+    maximumFractionDigits: 6,
+  }).format(value)
+}
+
 export function isSameSessionTree(
   current: string,
   sessionID: string,
